requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
git-filter-repo==2.47.0
google-auth==2.40.1
idna==3.10
kombu==5.5.3
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-certifi-win32==1.6.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-http-client==3.3.7
python3-openid==3.2.0
redis==6.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sendgrid==6.12.2
setuptools==80.7.1
setuptools-scm==8.3.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
